code cleanup and add tests
