+++ release/src/types.ts
@@ -41,7 +41,7 @@ export type Issue = {
   number: number;
   title: string;
   html_url: string;
-  labels: string | { name?: string }[];
+  labels: string | ({ name?: string }[]);
   assignee: null |  { login: string };
   created_at: string;
 };
